Anagram
Wow, that was kinda tough for a challenge ranked easy, but bubble sort
comes in for the save again. I’m not sure if this was the most
efficient method, but I feel like it did a great job following the
parameters. I’d like to find a more efficient manner if possible.